+++ apps/api-service/src/products/finnews/dto/finnews-summarize.request.dto.ts
@@ -1,0 +1,5 @@
+interface FinNewsSummarizerDto {
+  title: string | null | undefined
+  description: string | null | undefined
+  content: string | null | undefined
+}