+++ apps/ui-service/src/app/(platform)/products/wealthanalyzer/institution/[id]/page.tsx
@@ -14,7 +14,7 @@ import ky from "ky"
 import { uiConstants } from "@/shared/constants/global-constants"
 import notify from "@/shared/hooks/use-notify"
 import IconContainer from "@/shared/components/icon-container"
-import Summarizer from "@/shared/components/summarizer"
+import Summarizer from "@/shared/components/summarizer/asset-summarizer"
 
 export default function Page({ params }: { params: Promise<{ id: string }> }) {
   const { id: institutionId = "" } = use(params)