+++ apps/ui-service/src/app/layout.tsx
@@ -21,7 +21,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
       </head>
       <body className={quickSand.className}>
         <Providers>
-          <NextTopLoader color="#70f468" showSpinner={false} height={2} />
+          <NextTopLoader color="#1ff675" showSpinner={false} height={2} />
           <main className="min-h-screen w-full bg-main">{children}</main>
         </Providers>
       </body>