+++ apps/ui-service/src/shared/components/assetmodal/index.tsx
@@ -19,7 +19,7 @@ import { useConfirmContext } from "@/shared/providers/confirm.provider"
 import { useRouter } from "nextjs-toploader/app"
 import { Button } from "../ui/button"
 import { useUserContext } from "@/context/user.provider"
-import Summarizer from "../summarizer"
+import Summarizer from "../summarizer/asset-summarizer"
 import { useQueryClient } from "@tanstack/react-query"
 
 interface AssetModalProps {